[Refactor] #2444 Renamed apply-magic.* to item-magic-applier.*
