+++ src/floor/fixed-map-generator.cpp
@@ -18,8 +18,8 @@
 #include "monster-race/race-flags7.h"
 #include "monster/monster-util.h"
 #include "monster/smart-learn-types.h"
-#include "object-enchant/apply-magic.h"
 #include "object-enchant/item-apply-magic.h"
+#include "object-enchant/item-magic-applier.h"
 #include "object-enchant/object-ego.h"
 #include "object-enchant/trg-types.h"
 #include "object/object-info.h"
