+++ src/mind/mind-archer.cpp
@@ -11,8 +11,8 @@
 #include "grid/grid.h"
 #include "inventory/inventory-object.h"
 #include "io/command-repeater.h"
-#include "object-enchant/apply-magic.h"
 #include "object-enchant/item-apply-magic.h"
+#include "object-enchant/item-magic-applier.h"
 #include "object-enchant/object-boost.h"
 #include "object/item-tester-hooker.h"
 #include "object/item-use-flags.h"
