+++ src/object-enchant/item-magic-applier.h
