+++ src/store/store-util.cpp
@@ -5,9 +5,9 @@
  */
 
 #include "store/store-util.h"
-#include "object-enchant/apply-magic.h"
 #include "object-enchant/item-apply-magic.h"
 #include "object-enchant/item-feeling.h"
+#include "object-enchant/item-magic-applier.h"
 #include "object-enchant/special-object-flags.h"
 #include "object/object-kind.h"
 #include "object/object-value.h"
