+++ src/wizard/wizard-special-process.cpp
@@ -50,8 +50,8 @@
 #include "monster/monster-status.h"
 #include "monster/smart-learn-types.h"
 #include "mutation/mutation-investor-remover.h"
-#include "object-enchant/apply-magic.h"
 #include "object-enchant/item-apply-magic.h"
+#include "object-enchant/item-magic-applier.h"
 #include "object-enchant/trc-types.h"
 #include "object-enchant/trg-types.h"
 #include "object/object-kind-hook.h"
